add check isAuthenticated in controller

+++ src/controllers/post/DeletePostController.php
@@ -8,6 +8,7 @@ use App\lib\DatabaseConnection;
 use App\lib\PostIdChecker;
 use App\lib\SessionChecker;
 use App\lib\SessionManager;
+use App\Lib\UserChecker;
 use App\lib\View;
 use App\model\PostRepository;
 use Exception;
@@ -32,13 +33,21 @@ class DeletePostController
 
         $sessionChecker->sessionChecker();
         $sessionData = $sessionChecker->getSessionData();
+        $userChecker = new UserChecker();
+        $error = null;
 
         $view = new View();
         $postId = PostIdChecker::getId($args);
         $post = $this->getPostsRepository()->getPost($postId);
 
-        $html = $view->render('post_delete.twig', ['post' => $post, 'session' => $sessionData]);
-        $response->getBody()->write($html);
+        if (!$userChecker->isAuthenticated($sessionData['token'] ?? '')) {
+            $error = 'Vous n\'avez pas accès à cette page !';
+            $html = $view->render('error.twig', ['error' => $error]);
+            $response->getBody()->write($html);
+        } else {
+            $html = $view->render('post_delete.twig', ['post' => $post, 'session' => $sessionData, 'error' => $error]);
+            $response->getBody()->write($html);
+        }
 
         return $response;
     }
@@ -54,11 +63,28 @@ class DeletePostController
      */
     public function remove(RequestInterface $request, ResponseInterface $response, array $args): ResponseInterface
     {
-        if ($request->getMethod() === 'POST') {
-            $postRepository = $this->getPostsRepository();
-            $postId = PostIdChecker::getId($args);
-            $error = null;
+        $sessionManager = new SessionManager();
+        $sessionChecker = new SessionChecker($sessionManager);
 
+        $sessionChecker->sessionChecker();
+        $sessionData = $sessionChecker->getSessionData();
+        $userChecker = new UserChecker();
+
+        $postRepository = $this->getPostsRepository();
+        $postId = PostIdChecker::getId($args);
+        $fetchPost = $postRepository->getPost($postId);
+
+        $error = null;
+        if (
+            !($userChecker->isAuthenticated($sessionData['token'] ?? '') &&
+                $userChecker->isCurrentUser($fetchPost->userId, $sessionData['id']) ||
+                $userChecker->isAdmin($sessionData['role']))
+        ) {
+            $error = 'Vous n\'avez pas accès à cette page !';
+            return $this->renderErrorResponse($response, $error);
+        }
+
+        if ($request->getMethod() === 'POST') {
             $success = $postRepository->deletePost($postId);
 
             if ($success === false) {
@@ -74,8 +100,21 @@ class DeletePostController
 
             return $response;
         }
+    }
 
-        throw new Exception('Une erreur est survenue');
+    /**
+     * renderErrorResponse
+     *
+     * @param  ResponseInterface $response
+     * @param  string $error
+     * @return ResponseInterface
+     */
+    private function renderErrorResponse(ResponseInterface $response, string $error): ResponseInterface
+    {
+        $view = new View();
+        $html = $view->render('error.twig', ['error' => $error]);
+        $response->getBody()->write($html);
+        return $response;
     }
 
     /**
